pyproject.toml
[tool.poetry]
name = "real-estate"
version = "0.1.0"
description = ""
authors = ["Klim Belousov <klsbeousov@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.3"
requests = "2.31.0"
python-dotenv = "1.0.0"
psycopg2-binary = "2.9.9"
numpy = "1.26.4"
pandas = "2.2.2"
matplotlib = "3.8.4"
seaborn = "0.13.2"
scikit-learn = "1.4.2"
xgboost = "2.0.3"
lightgbm = "4.3.0"
catboost = "1.2.5"
dvc-s3 = "3.2.0"
dvc = "3.50.1"
mlflow = "2.12.2"
apache-airflow = "2.9.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
